Update key stock analysis.py
Amended some functions but the meat of the update was adding the finder function that I was using to scrape the data from the front page of a given ticker on yahoo finance. It was fundamentally a exercise in basic regex as well as iterating through data and lists to create dictionaries (conversion to JSON). 

I did find the use case to be extremely limited especially considering the amount of code it is, this was not by design, after some experimenting with selenium web driver i found the .text() function cleaned everything up much easier than I was able to. I'm still choosing to include it because I think it helps break down some fundamentals of regex, .find(), and splicing.